Add broken torrent download implementation
I think I'm pretty close to cracking it, but the interface between
anacrolix/torrent and the bubbletea Update() method doesn't quite work
yet. Currently the torrent gets added properly to the client, but it
quickly experiences an error that writes over the drawn content and
causes rendering problems thereafter. I will look into this and fix the
bug in future commits, but I wanted to get this basic foundation frozen
so I can mess around with impunity.