Fix critical bug by switching to MMap for storage
Using the default (which seemed to be SQLite?) caused a major glitch
after adding the torrent. Switching to MMap seems to have fixed this. I
have not tested this with the other available storage methods.

I have additionally moved the default storage to the current current
directory (rather than the OS temp directory). I'm looking into ways to
customize the download directory per-torrent.