Allow customization of torrent download dirs
This adds an optional "Save Directory" input to the add torrent prompt.
As part of this, torrents are no longer added directly to the client,
but instead converted to TorrentSpecs first to allow for customization
of the storage implementation per torrent.

For more documentation on TorrentSpec and its uses, see:
1. https://pkg.go.dev/github.com/anacrolix/torrent#TorrentSpec
2. https://github.com/anacrolix/torrent/discussions/584#discussioncomment-1303966