Add torrent metainfo to torrent details view
This commit adds the CreationDate, Comment, and CreatedBy fields of
torrent metainfo to the torrent details view when available. These
fields are rendered three newlines below the name of the torrent, and
above the download metadata (progress bar, files and peers, and download
and seeding stats).

These fields are optional, and do not exist on all torrents. They are
additionally only available when downloading torrents from a metainfo
(.torrent) file.

See BEP3 of the BitTorrent Protocol Specification for more information
on the metainfo dictionary.