Cap height of file view at 1/4 screen height
Previously, torrents with many files would cause rendering glitches as
they extended too far down and conflicted with other UI elements.

Now, the file view extends up until it hits 1/4 of the screen height.
Any files that cannot be rendered are enumerated, and the user is
notified on the final line of the file view of the truncated files.