Redesign errorView UI
This commit includes a number of changes to the styling and display of
errors, which are outlined below.

The app and popup borders have changed from the primary color (a light
blue) to the new error color (a light red). This change is quite
noticeable to the user, and conveys that something has gone wrong.
Further, the background character has been changed to draw a visual
distinction from the "adding torrent" screen.

The popup itself has undergone a number of changes. It is now 1/3 the
screen width wide, and 1/4 the screen height tall. It has horizontal
padding of 1/16 the screen width on both sides.

Textually, the header is no longer underlined and there is no line break
between it and the error content. Both the header and error are now
centered in the popup.

Finally, a tooltip has been added to guide the user to dismiss the
error.