Update dependencies to fix excess logging
anacrolix/go-libutp, a dependency of anacrolix/torrent, did not respect
the logger we configured for our torrent client and would log over the
Bubble Tea UI and caused errors in rendering for the rest of the session
until that section of the screen was refreshed.

We fixed these issues upstream, and this commit incorporates these
changes into Mabel.

Links:
  https://github.com/anacrolix/torrent/issues/676
  https://github.com/anacrolix/go-libutp/pull/19
  https://github.com/anacrolix/torrent/pull/722