Center main view
This commit moves the "You have no torrents!" prompt, displayed when the
client has no torrents, to the center of the screen. It also removes the
underline underneath the sentence, which was the last remnant of the
previous list design.

As a byproduct of this centering, the torrent list is now centered
vertically as well.