Make commit info optional in version output
This change makes it so that if the vcs.revision setting is empty and no
value is provided for the commit variable with ldflags, the version info
will no longer have an empty "Commit: " line. This brings commit info in
line with other types of version information like the build date.